Relocate traits (#2236)
* [breakpoints:debug] Rename service

* [breakpoints:debug] Add end new line.

* [breakpoints:debug] Apply PSR2 code style.

* [console] Relocate traits.

* [console] Validate commands using ContainerAwareCommandTrait.

* [console] Sync and fix conflicts.

* [console] Update command template.

* [console] Fix drupal coding style.
