+++ src/Command/ListCommand.php
@@ -12,6 +12,7 @@ use Symfony\Component\Console\Input\InputOption;
 use Drupal\Console\Helper\DescriptorHelper;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Command\Command as BaseCommand;
+use Drupal\Console\Command\Shared\CommandTrait;
 use Drupal\Console\Style\DrupalStyle;
 
 class ListCommand extends BaseCommand
